Factory markers now inherit constant-combinator
Previously when you wanted to put a sign on a factory building, there
was a special chest which you'd put an item in. This is problematic for
a few reasons: you're limited to only marking factories with item icons
(as opposed to signal icons); you need to actually get the item and use
it; and it won't play nice with blueprintability. This patch changes the
factory-overlay-controller to work like a constant combinator, so you
can choose any signal icon and don't need an item. It does not yet
handle blueprinting, but makes blueprinting feasible.